// bio-energy-frontend/src/models/IMeasure.ts
export interface PowerMeasure {
    id: string;
    unit: string | null;
    description: string | null;
  }
  export interface PowerMeasureDTO {
    unit: string| null;
    description: string| null;
  }
  export interface HeatMeasure {
    id: string;
    unit: string | null;
    description: string | null;
  }
  export interface HeatMeasureDTO {
    unit: string| null;
    description: string| null;
  }